charms/katib-db-manager/requirements-unit.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements-unit.in
#
anyio==3.6.2
    # via
    #   -r ./requirements.txt
    #   httpcore
certifi==2022.12.7
    # via
    #   -r ./requirements.txt
    #   httpcore
    #   httpx
charmed-kubeflow-chisme==0.0.8
    # via -r ./requirements.txt
coverage==7.2.3
    # via -r ./requirements-unit.in
deepdiff==6.2.1
    # via
    #   -r ./requirements.txt
    #   charmed-kubeflow-chisme
exceptiongroup==1.1.1
    # via pytest
h11==0.14.0
    # via
    #   -r ./requirements.txt
    #   httpcore
httpcore==0.16.3
    # via
    #   -r ./requirements.txt
    #   httpx
httpx==0.23.3
    # via
    #   -r ./requirements.txt
    #   lightkube
idna==3.4
    # via
    #   -r ./requirements.txt
    #   anyio
    #   rfc3986
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r ./requirements.txt
    #   charmed-kubeflow-chisme
lightkube==0.12.0
    # via
    #   -r ./requirements.txt
    #   charmed-kubeflow-chisme
lightkube-models==1.25.4.4
    # via
    #   -r ./requirements.txt
    #   lightkube
markupsafe==2.1.2
    # via
    #   -r ./requirements.txt
    #   jinja2
oci-image==1.0.0
    # via -r ./requirements.txt
ops==2.2.0
    # via
    #   -r ./requirements.txt
    #   charmed-kubeflow-chisme
ordered-set==4.1.0
    # via
    #   -r ./requirements.txt
    #   deepdiff
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
pytest==7.3.0
    # via
    #   -r ./requirements-unit.in
    #   pytest-mock
pytest-mock==3.10.0
    # via -r ./requirements-unit.in
pyyaml==6.0
    # via
    #   -r ./requirements.txt
    #   lightkube
    #   ops
rfc3986[idna2008]==1.5.0
    # via
    #   -r ./requirements.txt
    #   httpx
ruamel-yaml==0.17.21
    # via
    #   -r ./requirements.txt
    #   charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via
    #   -r ./requirements.txt
    #   ruamel-yaml
sniffio==1.3.0
    # via
    #   -r ./requirements.txt
    #   anyio
    #   httpcore
    #   httpx
tomli==2.0.1
    # via pytest
websocket-client==1.5.1
    # via
    #   -r ./requirements.txt
    #   ops
